requirements.txt
Django==3.1.3
gunicorn==20.0.4
django-heroku==0.3.1
requests==2.23.0
django_thumbs==0.4
django-ckeditor==6.0.0
python-decouple==3.4
